Migrate `common.param` (and friends) to return `BigInt` (#1854)
* Common: Return BigInt for param() functionality
Fixes #1846

* Change param getters to throw on non-existent value

* Fixes

* switch return to bigint or undefined

* Switch param getters to return undefined and fix common tests

* tx: changes to accomodate param getters

* More changes

* More adjustments

* Fix NaN bug in opcode fee builder

* Hack to fix EIP-2537 edge case

* client: fix test syntax

* lint

* Return 0 when param doesn't exist

* Spelling, dedup param usage

* Update `paramByEip` to allow undefined

* Hardcode EIP-2537 gas discount array

* uncomment test

+++ packages/vm/src/evm/precompiles/04-identity.ts
@@ -7,9 +7,8 @@ export default function (opts: PrecompileInput): ExecResult {
 
   const data = opts.data
 
-  let gasUsed = BigInt(opts._common.param('gasPrices', 'identity'))
-  gasUsed +=
-    BigInt(opts._common.param('gasPrices', 'identityWord')) * BigInt(Math.ceil(data.length / 32))
+  let gasUsed = opts._common.param('gasPrices', 'identity')
+  gasUsed += opts._common.param('gasPrices', 'identityWord') * BigInt(Math.ceil(data.length / 32))
 
   if (opts.gasLimit < gasUsed) {
     return OOGResult(opts.gasLimit)
